Move polynomial to libhelfem
