+++ libhelfem/CMakeLists.txt
@@ -8,6 +8,7 @@ add_library(helfem STATIC
     src/utils.cc
     src/chebyshev.cc
     src/lobatto.cc
+    src/polynomial.cc
 )
 target_include_directories(helfem PUBLIC "${CMAKE_CURRENT_BINARY_DIR}/include/" "${ARMADILLO_INCLUDE_DIRS}")
 target_include_directories(helfem PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/src/")
