+++ libhelfem/src/polynomial.cc
