+++ libhelfem/src/polynomial.h
