+++ src/atomic/quadrature.cpp
@@ -16,7 +16,7 @@
 #include "quadrature.h"
 #include "erfc_expn.h"
 #include "chebyshev.h"
-#include "../general/polynomial.h"
+#include "polynomial.h"
 #include "../general/utils.h"
 
 namespace helfem {
