+++ src/diatomic/inttest.cpp
@@ -14,7 +14,7 @@
  * of the License, or (at your option) any later version.
  */
 #include "quadrature.h"
-#include "../general/polynomial.h"
+#include "polynomial.h"
 #include "chebyshev.h"
 #include "../general/gaunt.h"
 #include "../general/utils.h"
