+++ src/harmonic/main.cpp
@@ -13,7 +13,7 @@
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
  */
-#include "../general/polynomial.h"
+#include "polynomial.h"
 #include "../general/polynomial_basis.h"
 #include "chebyshev.h"
 #include "quadrature.h"
