+++ src/sadatom/basis.cpp
@@ -14,7 +14,7 @@
  * of the License, or (at your option) any later version.
  */
 #include "basis.h"
-#include "../general/polynomial.h"
+#include "polynomial.h"
 #include "chebyshev.h"
 #include "../general/spherical_harmonics.h"
 #include "../general/gaunt.h"
